internet_shop-4_backend/pyproject.toml
[tool.poetry]
name = "internet-shop-4-backend"
version = "0.1.0"
description = ""
authors = ["Oleksandr Kondratiuk <alex_kondr@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
flask = "^3.1.0"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"
requests-html = "^0.10.0"
python-dotenv = "^1.0.1"
flask-migrate = "^4.1.0"
psycopg2 = "^2.9.10"
lxml = {extras = ["html-clean"], version = "^5.3.0"}
flask-jwt-extended = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
